requirements.txt
# Web2Ebook Requirements
# Copyright (c) devsimsek

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# EPUB generation
ebooklib>=0.18

# PDF generation
reportlab>=4.0.0
Pillow>=10.0.0

# HTML to text conversion
html2text>=2020.1.16

# Optional: For MOBI conversion (requires Calibre installed separately)
# Calibre's ebook-convert command-line tool must be installed